Enhance filter precision with 8-bit fixed-point math
The internal filter state (`_value`) now uses `uint16_t` to store values with an additional 8 bits of fractional precision (Q8.8 format).

This change mitigates precision loss that could occur with `uint8_t` when input deltas are small or smoothing factors are high, resulting in a significantly smoother filter output.

Additionally, the interpretation of the `_smoothing` parameter has been adjusted such that a higher value now corresponds to a stronger smoothing effect (less responsive to change), aligning with common filter terminology.

+++ src/dv_iir_filter8.h
@@ -15,7 +15,7 @@ public:
 
 private:
   uint8_t _smoothing = 32;
-  uint8_t _value;
+  uint16_t _value;
   bool _initialized = false;
   unsigned long _count = 0;
   static uint8_t lerp(uint8_t a, uint8_t b, uint8_t smoothing);
